상품목록 조회 , 상세 조회 완료

+++ IotPJ/src/main/java/com/iot/web/domain/ProductInfoDTO.java
@@ -10,10 +10,10 @@ import lombok.ToString;
 @Setter
 @ToString 
 public class ProductInfoDTO {
-    private Long productId;
+    private Integer productId;
     private String sellerId;
     private String productName;
-    private Long price;
+    private Integer price;
     private Double minTemperature;
     private Double maxTemperature;
     private Integer minHumidity;
